+++ yggdrasil-core/src/games/mod.rs
@@ -1,6 +1,8 @@
 pub mod invaders;
+pub mod poker;
 pub mod snake;
 pub mod tetris;
 pub use invaders::YggInvaders;
+pub use poker::YggPoker;
 pub use snake::{YggGame, YggSnake};
 pub use tetris::YggTetris;